requirements.txt
astroid==1.6.1
backports.functools-lru-cache==1.4
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.4
boto3==1.5.25
botocore==1.8.39
cachetools==2.0.1
certifi==2017.7.27.1
cffi==1.11.4
chardet==3.0.4
click==6.7
colorama==0.3.9
configparser==3.5.0
cycler==0.10.0
dj-database-url==0.4.2
Django==1.11.5
django-localflavor==2.0
django-storages==1.6.5
docutils==0.14
enum34==1.1.6
flake8==3.5.0
functools32==3.2.3.post2
futures==3.2.0
gunicorn==19.7.1
httplib2==0.10.3
idna==2.6
isort==4.2.15
itsdangerous==0.24
Jinja2==2.9.6
jmespath==0.9.3
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.13.3
oauth2client==4.1.2
pathlib==1.0.1
pew==1.1.2
Pillow==5.0.0
pipenv==9.0.3
postgres==2.2.1
psutil==5.3.1
psycopg2==2.7.3.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
pylint==1.7.4
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.12
shutilwhich==1.1.0
singledispatch==3.4.0.3
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
Werkzeug==0.12.2
whitenoise==3.3.1
wrapt==1.10.11
